packages/paper_elements/src/web-animations-next/test/blink/get-css-players.html

<!DOCTYPE html><script src="testharness/testharness.js"></script><script src="testharness/testharnessreport.js"></script><style>@keyframes test{from{opacity:0}to{opacity:1}}.cssAnimation{animation:test 2s}</style><div id=container><div id=element></div></div><script>async_test(function(a){assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0);element.className="cssAnimation";onload=function(){a.step(function(){var b=document.timeline.getAnimationPlayers();assert_equals(b.length,1);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,1);b[0].finish();assert_equals(document.timeline.getAnimationPlayers().length,0);assert_equals(container.getAnimationPlayers().length,0);assert_equals(element.getAnimationPlayers().length,0);a.done()})}},"getAnimationPlayers() with cssanimations");</script>
